functions/fbc-slackbot-python/requirements.txt
emoji_data_python
python-dotenv
slackclient
twilio
